Battle

See merge request s08-final/S08P31E103!432

+++ backend/battle/src/main/java/com/paymong/battle/battle/vo/common/MongStats.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class MongStats {
     private Long mongId;
     private Double health;
-    private Integer strength;
-    private Integer weight;
+    private Double strength;
+    private Double weight;
 }
